[music] Fix LE/gcc build: add pixdesc include + cast aspect to double
Two issues caught by LE's gcc with -Werror flags that MSVC didn't surface:

  1. MusicCodecInfoFFmpeg.cpp: av_color_transfer_name needs an explicit
     include of <libavutil/pixdesc.h>; cores/FFmpeg.h didn't pull it in
     transitively.

  2. MusicDatabase.cpp: passing a float to PrepareSQL's %f format trips
     -Werror=double-promotion since variadic args promote. Explicit cast
     to double in the video-row INSERT.

+++ xbmc/music/MusicDatabase.cpp
@@ -1579,7 +1579,7 @@ bool CMusicDatabase::SetStreamDetailsForSong(int idSong,
           " VALUES (%i, %i, 0, 0, '%s', %i, %i, %f, %i, '%s', '%s', '%s', '%s')",
           idSong, idAlbum,
           videoStream.strVideoCodec.c_str(), videoStream.iVideoWidth,
-          videoStream.iVideoHeight, videoStream.fVideoAspect,
+          videoStream.iVideoHeight, static_cast<double>(videoStream.fVideoAspect),
           videoStream.iVideoDuration, videoStream.strStereoMode.c_str(),
           videoStream.strVideoLanguage.c_str(), videoStream.strHdrType.c_str(),
           videoStream.strHdrDetail.c_str());
